Added flag to load CBA settings from file
Quote from https://github.com/CBATeam/CBA_A3/wiki/CBA-Settings-System

The cba_settings.sqf file of a mission is ignored by default. If the file is present when the mission is saved in the editor, a flag will be set and enable the mission to read the file. This only works if CBA is present when the mission is saved. To avoid creating a dependency, the mission maker can manually add the flag to the description.ext:
cba_settings_hasSettingsFile = 1;
Note that this will crash the mission on a dedicated server if the file is not present in the mission.